+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-
+### Fixed
+- getFilteredValue() without setting a Sanitizer did now throw the correct
+  Exception (namespacing typo)
 
 ## [1.1.0] - 2016-01-27
 ### Added